Merge in ONE/products-exp-service from hotfix/fixed-issue to develop

* commit '11d903456c270af8a0f94bcf9cd59ca616348b2d':
  fixed issue calculate range income invalid
  append log

+++ src/test/java/com/tmb/oneapp/productsexpservice/service/LoanSubmissionCustomerServiceTest.java
@@ -89,7 +89,7 @@ public class LoanSubmissionCustomerServiceTest {
         TmbOneServiceResponse<List<LoanOnlineRangeIncome>> oneServiceResponse = new TmbOneServiceResponse<>();
         List<LoanOnlineRangeIncome> rangeIncomeList = new ArrayList<>();
         LoanOnlineRangeIncome rangeIncome = new LoanOnlineRangeIncome();
-        rangeIncome.setRevenueMultiple(5);
+        rangeIncome.setRevenueMultiple(BigDecimal.valueOf(5));
         rangeIncome.setProductCode("RC01");
         rangeIncome.setEmploymentStatus("salary");
         rangeIncome.setRangeIncomeMaz(99999);
